feat(workflow): lifecycle methods on canonical WorkflowTask + TaskStatus to autobot_shared (closes #7121, partial #6520)
Two coordinated moves so canonical WorkflowTask owns its execution
lifecycle without taking a backend dependency.

1. Move TaskStatus + Severity + Priority + LLMProvider + OperationOutcome
   + HealthStatus enums to autobot_shared/status_enums.py.
   autobot-backend/constants/status_enums.py becomes a re-export shim
   (15+ existing callers keep working unchanged). Closes #6520 partial:
   the canonical TaskStatus now lives in autobot_shared, so
   autobot_shared.workflow can use the enum without inverting the
   layering.

2. Move 9 lifecycle methods (start_execution, complete_execution,
   fail_execution, get_execution_time, can_retry, increment_retry,
   get_enhanced_inputs, to_completed_result, to_failed_result) from
   enhanced_orchestration.types.AgentTask to canonical WorkflowTask in
   autobot_shared/workflow/types.py. AgentTask collapses to a pure
   alias (`AgentTask = WorkflowTask`). Originally extracted under
   #372 (feature-envy reduction).

Variance fix
------------
Before: `WorkflowPlan.tasks: List[WorkflowTask]` would accept a plain
WorkflowTask (no lifecycle methods) or an AgentTask subclass (with
methods), creating a silent AttributeError risk if a caller iterated
expecting `task.start_execution()`.

After: every WorkflowTask instance — plain canonical, aliased
(workflow_templates/orchestration.types WorkflowStep), or formerly
subclassed (AgentTask) — has all 9 methods. No more variance gap.

Verification
------------
- 29/29 canonical workflow tests pass (20 existing + 9 new lifecycle)
- 57/57 enhanced_orchestration tests pass (no regression on AgentTask
  callers despite the alias collapse)
- TaskStatus identity check passes via both import paths
  (constants.status_enums.TaskStatus is autobot_shared.status_enums.TaskStatus)
- Smoke test: WorkflowTask, AgentTask, workflow_templates.WorkflowStep
  all expose the 9 lifecycle methods

Closes #7121 — lifecycle methods on canonical
Closes #6520 partial — TaskStatus moved to autobot_shared (other duplicate
  TaskStatus enums in a2a/types.py, utils/task_queue.py,
  enhanced_memory_manager_async.py, services/agent_analytics.py,
  services/knowledge/task_status_manager.py still need their own
  consolidation; tracked under #6520)
